*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3506 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/func_basic.c
@@ -1150,7 +1150,7 @@ Var *f_random( Var *v )
 {
 	v = v;
 
-	return vars_push( FLOAT_VAR, ( double ) random( ) / ( double ) RAND_MAX );
+	return vars_push( FLOAT_VAR, random( ) / ( double ) RAND_MAX );
 }
